17
MIGRACION DE LAS TABLAS TERMINADA

+++ BACKEND/src/main/java/com/asistencia/attendance_system/model/dto/PracticanteRequest.java
@@ -11,7 +11,7 @@ public class PracticanteRequest {
     private String apellido;
     private String documento;
     @JsonAlias({"idAgencia"})
-    private Long idSede;
+    private Integer idSede;
     @JsonAlias({"idArea", "id_area"})
     private Long idPuesto;
     // Canónico nuevo: idArea (alias de idPuesto para migración)
